Implement Htmlable to allow use of escaped Blade tags ({{ $var }})
Allows users to use `{{ Notification::all() }}` instead of having to
manually call `toHTML()` and output using unescaped Blade tags.
(`{!! Notification::all()->toHTML() !!}`)

This unfortunately increases the dependency of `illuminate/support` to a
minimum of `5.1.3`, since that is when `Htmlable` was added.
(https://github.com/laravel/framework/pull/9271)

There is also now a cap on the dependency to `<5.3` just in case they
change `Htmlable` with the 5.3 release.

Signed-off-by: Micheal Mand <micheal@kmdwebdesigns.com>